Create gitignore tree  - Add essential files for recreatig projects
Every subfolder follows different gitignore rules.

For vivado project: create the write.tcl script + top wrapper and constrains.
For sdk project: add only the .c file that contains the main. 
For HLS IP: add the tcl script and the sources to recreate the HLS project. 
For RTL IP: add the verilog sources and the vivado packaged ip (xml and stuff)